+++ src/__init__.py
@@ -1,0 +1,8 @@
+import malpaca
+
+
+def main():
+	malpaca.main()
+
+if __name__ == "__main__":
+	main()
